Move Metadata Container to src/app/components
