+++ src/app/pages/MostReadPage/MostReadPage.jsx
@@ -18,10 +18,10 @@ import ATIAnalytics from '#containers/ATIAnalytics';
 import LinkedData from '#containers/LinkedData';
 import ChartbeatAnalytics from '#containers/ChartbeatAnalytics';
 import ComscoreAnalytics from '#containers/ComscoreAnalytics';
-import MetadataContainer from '#containers/Metadata';
 import Grid, { GelPageGrid } from '#components/Grid';
 import IndexPageContainer from '#components/PageLayout/IndexPageContainer';
 import IndexHeading from '#containers/IndexHeading';
+import MetadataContainer from '../../components/Metadata';
 import { ServiceContext } from '../../contexts/ServiceContext';
 
 const MarginWrapper = styled.div`
