+++ src/app/pages/TopicPage/TopicPage.jsx
@@ -2,7 +2,6 @@ import React, { useContext } from 'react';
 import ATIAnalytics from '#containers/ATIAnalytics';
 import { shape, arrayOf, string } from 'prop-types';
 import path from 'ramda/src/path';
-import MetadataContainer from '#containers/Metadata';
 import LinkedData from '#containers/LinkedData';
 import AdContainer from '#containers/Ad';
 import CanonicalAdBootstrapJs from '#containers/Ad/Canonical/CanonicalAdBootstrapJs';
@@ -11,6 +10,7 @@ import { RequestContext } from '#contexts/RequestContext';
 import ChartbeatAnalytics from '#containers/ChartbeatAnalytics';
 import Curation from '#app/components/Curation';
 import styles from './index.styles';
+import MetadataContainer from '../../components/Metadata';
 import { ServiceContext } from '../../contexts/ServiceContext';
 import TopicImage from './TopicImage';
 import TopicTitle from './TopicTitle';
