+++ ws-nextjs-app/pages/[service]/new_live/[id]/LivePageLayout.tsx
@@ -3,12 +3,12 @@
 
 import React, { useContext } from 'react';
 import { jsx } from '@emotion/react';
-import MetadataContainer from '#app/legacy/containers/Metadata';
 import LinkedDataContainer from '#app/legacy/containers/LinkedData';
 import Pagination from '#pages/TopicPage/Pagination';
 import Heading from '#app/components/Heading';
 import { ServiceContext } from '#contexts/ServiceContext';
 import nodeLogger from '#lib/logger.node';
+import MetadataContainer from '../../../../../src/app/components/Metadata';
 
 import styles from './styles';
